refactor: Simplify Docker build command in README
The Docker build command in the README has been simplified to remove the `--no-cache` flag. This change aligns with the recent refactor of the Dockerfile and docker-entrypoint.sh script, which improved the logic for running the server and client. The simplified build command will still create the data-guild-grpc image for the project.