benchmarking code from Data guild (#4)
* bump to 0.1.1 : barebone grpc demo
ready for liftoff

* added compression on both client and server

* make ing life easier

* added server for benchmark

* added the client

* bugfix library name

* bugfix the delete benchmarking

* adding lib for rest api benchmark

* rest api benchmark with flask

* perform mutiple ops and average result

* calculate tranfer volume

* bugfix : control item id from client

* remove redundant import

* counting in ms
also using monotonic

* grpc : calculating byte transfers and transfer rate

* grpc : integrate typer

* rest client : integrate cli magic

* linter fix

* linter : doc string from copilot

* aligning rest server with client

* adding dockerfile for grpc server

* rename

* add entrypoint script

* feat: Add docker-entrypoint.sh script

The `docker-entrypoint.sh` script is added to the repository. This script serves as the entry point for the Docker container and handles different commands based on the provided argument. It includes logic for running the publisher, starting the subscriber, and executing unfamiliar commands.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature).

* refactor: Update docker-entrypoint.sh script

The `docker-entrypoint.sh` script has been updated to improve the logic for running the server and client. It now includes initialization of the database for the server and benchmarking for the client. This change aligns with the purpose of the script as the entry point for the Docker container

* ignore locally created db

* entrypoint with multiple command for image

* refactor: Update Dockerfile and docker-entrypoint.sh

The Dockerfile and docker-entrypoint.sh script have been updated to improve the logic for running the server and client. The changes include copying the project files, installing Poetry, and setting up the project dependencies. Additionally, the docker-entrypoint.sh script now has executable permissions. This refactor aligns with the purpose of the Dockerfile and script as the configuration and entry point for the Docker container.

* readme : instructions for docker

* refactor: Simplify Docker build command in README

The Docker build command in the README has been simplified to remove the `--no-cache` flag. This change aligns with the recent refactor of the Dockerfile and docker-entrypoint.sh script, which improved the logic for running the server and client. The simplified build command will still create the data-guild-grpc image for the project.

---------

Co-authored-by: Ratul <89194947+ratul-cefalo@users.noreply.github.com>